Diarize fixes
