+++ backend/.gitignore
@@ -6,3 +6,4 @@ app.egg-info
 htmlcov
 .cache
 .venv
+app/hf_token.py
